+++ src/Sulu/Bundle/PreviewBundle/Preview/Exception/WebspaceNotFoundException.php
@@ -12,15 +12,21 @@
 namespace Sulu\Bundle\PreviewBundle\Preview\Exception;
 
 /**
- * This exception will be thrown when no portal exists for given locale and webspace.
+ * This exception will be thrown when the given webspace does not exist.
  */
-class PortalNotFoundException extends PreviewRendererException
+class WebspaceNotFoundException extends PreviewRendererException
 {
+    /**
+     * @param object $object
+     * @param int $id
+     * @param string $webspaceKey
+     * @param string $locale
+     */
     public function __construct($object, $id, $webspaceKey, $locale)
     {
         parent::__construct(
-            sprintf('Portal for combination of webspace "%s" and locale "%s" not found.', $webspaceKey, $locale),
-            self::BASE_CODE + 1,
+            sprintf('Webspace "%s" not found', $webspaceKey),
+            self::BASE_CODE + 6,
             $object,
             $id,
             $webspaceKey,